% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_data}
\alias{dummy_data}
\title{createdummy}
\usage{
dummy_data(nbcol, nbrow, VarName = paste0("var_", 1:nbcol))
}
\arguments{
\item{nbcol}{integer, the number of columns (i.e. variables) in your dummy data}

\item{nbrow}{integer, the number of rows (i.e. observations) in your dummy data}

\item{VarName}{character vector of length ncol, the name of your variables}
}
\value{
a dummy data frame
}
\description{
create a dummy numeric data frame
}
