% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_legend.R
\name{extract_legend}
\alias{extract_legend}
\title{Title}
\usage{
extract_legend(my_ggp)
}
\arguments{
\item{my_ggp}{a ggplot object to which you want to get the legend}
}
\value{
a legend object that can be added to a ggplot
}
\description{
Title
}
