% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotIn.R
\name{\%notin\%}
\alias{\%notin\%}
\title{NotIn}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector to be tested}

\item{y}{vectore to be tested against}
}
\value{
notin operator
}
\description{
create an operator being the opposite of \%in\%
}
