% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIRAS.R
\name{merge_ciras}
\alias{merge_ciras}
\title{merge_ciras}
\usage{
merge_ciras(
  path_to_xls = "C:/Users/trist/AppData/Local/ProjetsR/Greenhouse_Holobrom/Ciras/clean"
)
}
\arguments{
\item{path_to_xls}{a character string with your path to all your ciras xls output.
their name must always end as _treatment_sampleID.xls (e.g. CIRAS_3_Aechmea m _DP_1.xls).}
}
\value{
csv individual and merged frames. Full and reducted to relevant variables.
}
\description{
get all xls ciras output from a directory and transform them into
csv in a 'csv' subfolder. Additionally bind them into to files (one complete, one
with only relevant variables) in a 'final' subfolder.
}
