% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_minipam.R
\name{merge_minipam}
\alias{merge_minipam}
\title{merge_minipam}
\usage{
merge_minipam(input_path)
}
\arguments{
\item{input_path}{character, the path to your minipam csv output to be merged}
}
\value{
a csv, row binding all the files contained it your input path
}
\description{
merge_minipam
}
