% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIPAM.R
\name{minipam}
\alias{minipam}
\title{minipam}
\usage{
minipam(
  Name_of_the_input_file,
  input_path = "C:/Users/trist/AppData/Local/ProjetsR/Greenhouse_Holobrom/MiniPam/data",
 
    output_path = "C:/Users/trist/AppData/Local/ProjetsR/Greenhouse_Holobrom/MiniPam/data/graph"
)
}
\arguments{
\item{Name_of_the_input_file}{a character string (or vector), name of the csv input.}

\item{input_path}{the path leading to the input file.}

\item{output_path}{the path where you want to have the graphs stored.}
}
\value{
dataframes and graphs.
}
\description{
take as input a csv dataframe containing output
of the minipam. For ETR and FvFm measurements only. And return
clean files containing each type of measurements + the ETR curves.
}
