% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PASCO_transfo2.R
\name{PASCO_transfo2}
\alias{PASCO_transfo2}
\title{Pasco_transfo2}
\usage{
PASCO_transfo2(data, ech, name_run, select, A = 1, V = 5)
}
\arguments{
\item{data}{a data frame output from Sparkview (usually read from .csv)}

\item{ech}{a character vector with either the probe or sample name}

\item{name_run}{a character vector with the name of all your runs (e.g., c("stab1","RECO","NEE"))}

\item{select}{a numeric vector of the runs you want to keep (e.g., c(2,3))}

\item{A}{the Area}

\item{V}{the Volume}
}
\value{
a csv file prepared for gasfluxes package
}
\description{
Pasco_transfo2
}
