% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_layout.R
\name{plate_layout}
\alias{plate_layout}
\title{plate_layout}
\usage{
plate_layout(
  samples,
  proj,
  name_file,
  save_file_path,
  starting_plate_number = 1
)
}
\arguments{
\item{samples}{a vector containing all your samples ID, they will fill the plate in the order they are in this vector}

\item{proj}{name of your porject to name your plates as : "proj-PLx"}

\item{name_file}{a name to your output file}

\item{save_file_path}{path to where you want to save the excel output}

\item{starting_plate_number}{where from start plate numbering}
}
\value{
an xlsx workbook with you plate plan
}
\description{
plate_layout
}
