% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSYPRO.R
\name{psypro}
\alias{psypro}
\title{psypro}
\usage{
psypro(usedset, lim = c(-3, 2), ID_vec, path_to_calibration, psypro_output)
}
\arguments{
\item{usedset}{the predetermined name of your set 0,1,2 or 3.}

\item{lim}{min and max values expected out of the psypro for you
samples. Used to standardized graphs for faster reading. Discuss with
lab members to understand!!}

\item{ID_vec}{a vector of length 8 (number of sensors) with your samples' ID.
Empty sensors are named 0}

\item{path_to_calibration}{path to you calibration file.}

\item{psypro_output}{path to your psypro output}
}
\value{
dataframes and graphs.
}
\description{
Transform psypro output files into csv dataframe with
mean water potential of your triplicate.
}
