% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_layout.R
\name{tag_layout}
\alias{tag_layout}
\title{tag_layout}
\usage{
tag_layout(
  tag_list,
  PCR_plates,
  output_path,
  file_corresp_tag,
  file_tag_layout
)
}
\arguments{
\item{tag_list}{a dataframe with 3 column : 'tag_name' (e.g. f1 to fx and r1 to rx), 'tag_sequence' (e.g. ACACACAC) and 'tag_type' (i.e. forward or reverse)}

\item{PCR_plates}{a matrix object representing your plates map/layout, output of "plate_layout" function of this package.}

\item{output_path}{path to an output folder that will receive to new files}

\item{file_corresp_tag}{name of the sample-tagpairs correspondance dataframe}

\item{file_tag_layout}{name of your xlsx output, having the map of your tagz.}
}
\value{
an xlsx workbook with you tag plate plan and a csv file with your correspondance between tags and samples
}
\description{
tag_layout
}
