% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIRAS.R
\name{merge_ciras}
\alias{merge_ciras}
\title{merge_ciras}
\usage{
merge_ciras(path_to_xls, skip = 3)
}
\arguments{
\item{path_to_xls}{a character string with your path to all your ciras xls output.
their name must always end as _treatment_sampleID.xls (e.g. CIRAS_3_Aechmea m _DP_1.xls).}

\item{skip}{a integer indicating the number of useless rows at the top of your xls file, Jean-Yves Goret
actual template have three.}
}
\value{
csv individual and merged frames. Full and reducted to relevant variables.
}
\description{
get all xls ciras output from a directory and transform them into
csv in a 'csv' subfolder. Additionally bind them into to files (one complete, one
with only relevant variables) in a 'final' subfolder.
}
