% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIPAM.R
\name{minipam}
\alias{minipam}
\title{minipam}
\usage{
minipam(Name_of_the_input_file, input_path, rec_vec = NULL)
}
\arguments{
\item{Name_of_the_input_file}{a character string (or vector), name of the csv input (if possible the output of merge_minipam).}

\item{input_path}{the path leading to the input file.}

\item{rec_vec}{a vector containing the REC id of the measurements you want to keep in the final frame. Default is NULL, it will keep all the measures from the original file.}
}
\value{
dataframes and graphs.
}
\description{
take as input a csv dataframe containing output
of the minipam. For ETR and FvFm measurements only. And return
clean files containing each type of measurements + the ETR curves.
}
