% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIPAM.R
\name{minipam}
\alias{minipam}
\title{minipam}
\usage{
minipam(Name_of_the_input_file, input_path, path_to_ID_match)
}
\arguments{
\item{Name_of_the_input_file}{a character string (or vector), name of the csv input (if possible the output of merge_minipam).}

\item{input_path}{the path leading to the input file.}

\item{path_to_ID_match}{a csv file containing the measure ID, date, time and REC of the measure.}
}
\value{
dataframes and graphs.
}
\description{
take as input a csv dataframe containing output
of the minipam. For ETR and FvFm measurements only. And return
clean files containing each type of measurements + the ETR curves.
}
